Add files via upload
We worked through downloading a dictionary and putting it into a text file for Testa to use his Scanner method. We then created different methods for Testa's use, one being to determine if a word was inside our Dictionary. As well as another of determining if there was another word after the one passed to us.